Add files via upload
Before you read the changes, this is an unfinished pre alpha version. Meaning that there are some leftovers from tests n stuff.

New changes:
Drastical changes in the rendering include:
- Grass color
- Game colors and rendering, it now somewhat looks like Minecraft Beta 1.6.6/Console Edition TU1 and that's caused by an intentional change to the vignette.
- Changed vignette. Described above.
- Added all block textures from Java Edition including new 1.17 ores with almost the exact same file names(the lily pad texture file name differs)
- Better ambient occlusion on sand
- Better ambient occlusion on glowstone
- Fixed sand walking sounds.
- Nether portal trigger sound now matches Java Edition
- Water now matches Java Edition, but the water is greatly transparent and will be fixed very soon.
- Extra: Added Plastic Pack.

Trivia:
I was going to remove the post proccessing effect(color inversion) on the crosshair to match it to console edition, but when I did so, the crosshair didn't disappear when I turned off touch controls, I tried to fix it but the game simply doesn't load up. Shame.